+++ flutter_quill_extensions/lib/embeds/toolbar/video_button.dart
@@ -53,7 +53,7 @@ class VideoButton extends StatelessWidget {
     final iconFillColor =
         iconTheme?.iconUnselectedFillColor ?? (fillColor ?? theme.canvasColor);
 
-    return QuillIconButton(
+    return QuillToolbarIconButton(
       icon: Icon(icon, size: iconSize, color: iconColor),
       tooltip: tooltip,
       highlightElevation: 0,